fix otel settings for grpc
