+++ server/executor/trigger/instrument.go
@@ -2,9 +2,19 @@ package trigger
 
 import (
 	"context"
+	"io"
 
 	"github.com/kubeshop/tracetest/server/model"
+	"go.opentelemetry.io/contrib/propagators/aws/xray"
+	"go.opentelemetry.io/contrib/propagators/b3"
+	"go.opentelemetry.io/contrib/propagators/jaeger"
+	"go.opentelemetry.io/contrib/propagators/ot"
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace"
+	"go.opentelemetry.io/otel/propagation"
+	"go.opentelemetry.io/otel/sdk/resource"
+	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.4.0"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -49,4 +59,30 @@ func (t *instrumentedTriggerer) Trigger(ctx context.Context, test model.Test, ti
 	span.SetAttributes(attrs...)
 
 	return resp, err
+}
+
+func propagators() propagation.TextMapPropagator {
+	return propagation.NewCompositeTextMapPropagator(propagation.Baggage{},
+		b3.New(),
+		jaeger.Jaeger{},
+		ot.OT{},
+		xray.Propagator{},
+		propagation.TraceContext{})
+}
+
+func traceProvider() *sdktrace.TracerProvider {
+	// Set standard attributes per semantic conventions
+	res := resource.NewWithAttributes(
+		semconv.SchemaURL,
+		semconv.ServiceNameKey.String("tracetest"),
+	)
+
+	// this is in fact a noop exporter, so we can ignore errors
+	spanExporter, _ := stdouttrace.New(stdouttrace.WithWriter(io.Discard))
+
+	return sdktrace.NewTracerProvider(
+		sdktrace.WithSyncer(spanExporter),
+		sdktrace.WithResource(res),
+		sdktrace.WithSampler(sdktrace.ParentBased(sdktrace.AlwaysSample())),
+	)
 }
